Diretional Indicator Update
Indicator is now inactive when the game has started.
Ability to touch bell has also been removed. This is because the bell collider was too onconsistent in detecting touch, so to not annoy players, it was removed entirely.
I MIGHT add it later with a different idea... We will see...