Size increase for tile collider and entire meeple
This change was made AFTER the thesis work for public tryouts and demonstrations.
Collider on tile was increased as it was observed that my previous increase was not enough for a lot of people.
The size of the meeples were because a lot of people found the size of the game quite small, so now people may notice them more. They are also not too big in my opinion and still fit the tiles.